Set up map (#3)
The first task in the project is to render a map. The map has a fixed
size of 80 by 50 tiles, and uses ASCII characters for the different tile
types.

The project uses bracket-lib as the framework to build the game, which
is a very lightweight and easy-to-use library for beginners.