Add graphics and camera (#6)
The ASCII art has been replaced with pixel art, and the different
systems in the game have been updated to support that. The game now
uses a camera to render a section of the map, as well as layers to
render the player on top of the map.

Credits for the pixel art can be found in the relevant section of the
book.