Refactor game to use an ECS (#7)
Entity Component Systems (ECS) are a popular approach for managing
complexity in games. Entities have components, which describe their
abilities. Systems query these components, and perform actions based
on them. For example, a render system might query every entity with a
`Render` component and draw it on the screen.

The project uses Legion as the ECS, which has been added as a
dependency. The previous version of the game has been refactored to use
the ECS with a handful of systems, but no new functionality has been
added yet.