Implement victory and defeat (#11)
The game can now be won or lost. If the player loses all their health,
they die and have to restart. They can win the game by picking up the
Amulet of Yala, which is spawned as an item in the dungeon. Both the
victory and defeat states have their own screen, which informs the
player of the outcome and offers them to restart the game.