Add field of view (#12)
The rendering pipeline has been updated to take the entity's field of
view into consideration. The player can only see tiles in a certain
radius and within their line of sight, and monsters will only attack a
player if they can see them.

Tiles that a player has previously seen, but that are now outside of
their view, are rendered in gray.